Alterações Aula 297

+++ app/Http/Controllers/MarcaController.php
@@ -8,6 +8,10 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
+    public function __construct(Marca $marca){
+        $this->marca = $marca;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -16,7 +20,7 @@ class MarcaController extends Controller
     public function index()
     {
         //
-        $marcas = Marca::all();
+        $marcas = $this->marca->all();
         return $marcas;
     }
 
@@ -29,19 +33,23 @@ class MarcaController extends Controller
     public function store(Request $request)
     {
         //
-        $marca = Marca::create($request->all());
+        $marca = $this->marca->create($request->all());
         return $marca;
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function show(Marca $marca)
+    public function show($id)
     {
         //
+        $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro'=>'Recurso pesquisado não existe'];
+        }
         return $marca;
     }
 
@@ -49,12 +57,16 @@ class MarcaController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request, $id)
     {
         //
+        $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro'=>'Impossível realizar a atualização. O recurso solicitado não existe'];
+        }
         $marca->update($request->all());
         return $marca;
     }
@@ -62,12 +74,17 @@ class MarcaController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
         //
+        $marca = $this->marca->find($id);
+        $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro'=>'Impossível realizar a exclusão. O recurso solicitado não existe'];
+        }
         $marca->delete();
         return ['msg' => 'A marca foi removida com sucesso!'];
     }
